+++ core/services/ocr2/delegate.go
@@ -25,6 +25,7 @@ import (
 	ocr2keepers20runner "github.com/smartcontractkit/chainlink-automation/pkg/v2/runner"
 	ocr2keepers21config "github.com/smartcontractkit/chainlink-automation/pkg/v3/config"
 	ocr2keepers21 "github.com/smartcontractkit/chainlink-automation/pkg/v3/plugin"
+	"github.com/smartcontractkit/chainlink-common/pkg/loop/reportingplugins/ocr3"
 	"github.com/smartcontractkit/chainlink/v2/core/config/env"
 
 	"github.com/smartcontractkit/chainlink-vrf/altbn_128"
@@ -35,6 +36,7 @@ import (
 	"github.com/smartcontractkit/chainlink-common/pkg/loop"
 	"github.com/smartcontractkit/chainlink-common/pkg/loop/reportingplugins"
 	"github.com/smartcontractkit/chainlink-common/pkg/types"
+	llotypes "github.com/smartcontractkit/chainlink-common/pkg/types/llo"
 	"github.com/smartcontractkit/chainlink-common/pkg/utils/mailbox"
 
 	"github.com/smartcontractkit/chainlink/v2/core/bridges"
@@ -43,11 +45,14 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/logger"
 	"github.com/smartcontractkit/chainlink/v2/core/services/job"
 	"github.com/smartcontractkit/chainlink/v2/core/services/keystore"
+	"github.com/smartcontractkit/chainlink/v2/core/services/keystore/chaintype"
 	"github.com/smartcontractkit/chainlink/v2/core/services/keystore/keys/ocr2key"
+	"github.com/smartcontractkit/chainlink/v2/core/services/llo"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/dkg"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/dkg/persistence"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/functions"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/generic"
+	lloconfig "github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/llo/config"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/median"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/mercury"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ocr2keeper"
@@ -69,6 +74,7 @@ import (
 	evmmercury "github.com/smartcontractkit/chainlink/v2/core/services/relay/evm/mercury"
 	mercuryutils "github.com/smartcontractkit/chainlink/v2/core/services/relay/evm/mercury/utils"
 	evmrelaytypes "github.com/smartcontractkit/chainlink/v2/core/services/relay/evm/types"
+	"github.com/smartcontractkit/chainlink/v2/core/services/streams"
 	"github.com/smartcontractkit/chainlink/v2/core/services/synchronization"
 	"github.com/smartcontractkit/chainlink/v2/core/services/telemetry"
 	"github.com/smartcontractkit/chainlink/v2/plugins"
@@ -106,6 +112,7 @@ type Delegate struct {
 	bridgeORM             bridges.ORM
 	mercuryORM            evmmercury.ORM
 	pipelineRunner        pipeline.Runner
+	streamRegistry        streams.Getter
 	peerWrapper           *ocrcommon.SingletonPeerWrapper
 	monitoringEndpointGen telemetry.MonitoringEndpointGenerator
 	cfg                   DelegateConfig
@@ -118,7 +125,8 @@ type Delegate struct {
 	isNewlyCreatedJob bool // Set to true if this is a new job freshly added, false if job was present already on node boot.
 	mailMon           *mailbox.Monitor
 
-	legacyChains legacyevm.LegacyChainContainer // legacy: use relayers instead
+	legacyChains         legacyevm.LegacyChainContainer // legacy: use relayers instead
+	capabilitiesRegistry types.CapabilitiesRegistry
 }
 
 type DelegateConfig interface {
@@ -217,6 +225,7 @@ func NewDelegate(
 	bridgeORM bridges.ORM,
 	mercuryORM evmmercury.ORM,
 	pipelineRunner pipeline.Runner,
+	streamRegistry streams.Getter,
 	peerWrapper *ocrcommon.SingletonPeerWrapper,
 	monitoringEndpointGen telemetry.MonitoringEndpointGenerator,
 	legacyChains legacyevm.LegacyChainContainer,
@@ -228,6 +237,7 @@ func NewDelegate(
 	ethKs keystore.Eth,
 	relayers RelayGetter,
 	mailMon *mailbox.Monitor,
+	capabilitiesRegistry types.CapabilitiesRegistry,
 ) *Delegate {
 	return &Delegate{
 		db:                    db,
@@ -235,6 +245,7 @@ func NewDelegate(
 		bridgeORM:             bridgeORM,
 		mercuryORM:            mercuryORM,
 		pipelineRunner:        pipelineRunner,
+		streamRegistry:        streamRegistry,
 		peerWrapper:           peerWrapper,
 		monitoringEndpointGen: monitoringEndpointGen,
 		legacyChains:          legacyChains,
@@ -247,6 +258,7 @@ func NewDelegate(
 		RelayGetter:           relayers,
 		isNewlyCreatedJob:     false,
 		mailMon:               mailMon,
+		capabilitiesRegistry:  capabilitiesRegistry,
 	}
 }
 
@@ -431,6 +443,9 @@ func (d *Delegate) ServicesForSpec(jb job.Job) ([]job.ServiceCtx, error) {
 	case types.Mercury:
 		return d.newServicesMercury(ctx, lggr, jb, bootstrapPeers, kb, ocrDB, lc, ocrLogger)
 
+	case types.LLO:
+		return d.newServicesLLO(ctx, lggr, jb, bootstrapPeers, kb, ocrDB, lc, ocrLogger)
+
 	case types.Median:
 		return d.newServicesMedian(ctx, lggr, jb, bootstrapPeers, kb, ocrDB, lc, ocrLogger)
 
@@ -454,7 +469,7 @@ func (d *Delegate) ServicesForSpec(jb job.Job) ([]job.ServiceCtx, error) {
 		return d.newServicesOCR2Functions(lggr, jb, bootstrapPeers, kb, ocrDB, thresholdPluginDB, s4PluginDB, lc, ocrLogger)
 
 	case types.GenericPlugin:
-		return d.newServicesGenericPlugin(ctx, lggr, jb, bootstrapPeers, kb, ocrDB, lc, ocrLogger)
+		return d.newServicesGenericPlugin(ctx, lggr, jb, bootstrapPeers, kb, ocrDB, lc, ocrLogger, d.capabilitiesRegistry)
 
 	default:
 		return nil, errors.Errorf("plugin type %s not supported", spec.PluginType)
@@ -463,7 +478,7 @@ func (d *Delegate) ServicesForSpec(jb job.Job) ([]job.ServiceCtx, error) {
 
 func GetEVMEffectiveTransmitterID(jb *job.Job, chain legacyevm.Chain, lggr logger.SugaredLogger) (string, error) {
 	spec := jb.OCR2OracleSpec
-	if spec.PluginType == types.Mercury {
+	if spec.PluginType == types.Mercury || spec.PluginType == types.LLO {
 		return spec.TransmitterID.String, nil
 	}
 
@@ -490,12 +505,11 @@ func GetEVMEffectiveTransmitterID(jb *job.Job, chain legacyevm.Chain, lggr logge
 		effectiveTransmitterID, err := chain.TxManager().GetForwarderForEOA(common.HexToAddress(spec.TransmitterID.String))
 		if err == nil {
 			return effectiveTransmitterID.String(), nil
-		} else if spec.TransmitterID.Valid {
-			lggr.Warnw("Skipping forwarding for job, will fallback to default behavior", "job", jb.Name, "err", err)
-			// this shouldn't happen unless behaviour above was changed
-		} else {
+		} else if !spec.TransmitterID.Valid {
 			return "", errors.New("failed to get forwarder address and transmitterID is not set")
 		}
+		lggr.Warnw("Skipping forwarding for job, will fallback to default behavior", "job", jb.Name, "err", err)
+		// this shouldn't happen unless behaviour above was changed
 	}
 
 	return spec.TransmitterID.String, nil
@@ -514,38 +528,39 @@ func (d *Delegate) newServicesGenericPlugin(
 	ocrDB *db,
 	lc ocrtypes.LocalConfig,
 	ocrLogger commontypes.Logger,
+	capabilitiesRegistry types.CapabilitiesRegistry,
 ) (srvs []job.ServiceCtx, err error) {
 	spec := jb.OCR2OracleSpec
 
 	// NOTE: we don't need to validate this config, since that happens as part of creating the job.
 	// See: validate/validate.go's `validateSpec`.
-	p := validate.OCR2GenericPluginConfig{}
-	err = json.Unmarshal(spec.PluginConfig.Bytes(), &p)
+	pCfg := validate.OCR2GenericPluginConfig{}
+	err = json.Unmarshal(spec.PluginConfig.Bytes(), &pCfg)
 	if err != nil {
 		return nil, err
 	}
 
-	plugEnv := env.NewPlugin(p.PluginName)
+	plugEnv := env.NewPlugin(pCfg.PluginName)
 
-	command := p.Command
+	command := pCfg.Command
 	if command == "" {
 		command = plugEnv.Cmd.Get()
 	}
 
 	// Add the default pipeline to the pluginConfig
-	p.Pipelines = append(
-		p.Pipelines,
+	pCfg.Pipelines = append(
+		pCfg.Pipelines,
 		validate.PipelineSpec{Name: "__DEFAULT_PIPELINE__", Spec: jb.Pipeline.Source},
 	)
 
 	rid, err := spec.RelayID()
 	if err != nil {
-		return nil, ErrJobSpecNoRelayer{PluginName: p.PluginName, Err: err}
+		return nil, ErrJobSpecNoRelayer{PluginName: pCfg.PluginName, Err: err}
 	}
 
 	relayer, err := d.RelayGetter.Get(rid)
 	if err != nil {
-		return nil, ErrRelayNotEnabled{Err: err, Relay: spec.Relay, PluginName: p.PluginName}
+		return nil, ErrRelayNotEnabled{Err: err, Relay: spec.Relay, PluginName: pCfg.PluginName}
 	}
 
 	provider, err := relayer.NewPluginProvider(ctx, types.RelayArgs{
@@ -554,7 +569,7 @@ func (d *Delegate) newServicesGenericPlugin(
 		ContractID:    spec.ContractID,
 		New:           d.isNewlyCreatedJob,
 		RelayConfig:   spec.RelayConfig.Bytes(),
-		ProviderType:  p.ProviderType,
+		ProviderType:  pCfg.ProviderType,
 	}, types.PluginArgs{
 		TransmitterID: spec.TransmitterID.String,
 		PluginConfig:  spec.PluginConfig.Bytes(),
@@ -564,39 +579,19 @@ func (d *Delegate) newServicesGenericPlugin(
 	}
 	srvs = append(srvs, provider)
 
-	oracleEndpoint := d.monitoringEndpointGen.GenMonitoringEndpoint(
-		rid.Network,
-		rid.ChainID,
-		spec.ContractID,
-		synchronization.TelemetryType(p.TelemetryType),
-	)
-	oracleArgs := libocr2.OCR2OracleArgs{
-		BinaryNetworkEndpointFactory: d.peerWrapper.Peer2,
-		V2Bootstrappers:              bootstrapPeers,
-		Database:                     ocrDB,
-		LocalConfig:                  lc,
-		Logger:                       ocrLogger,
-		MonitoringEndpoint:           oracleEndpoint,
-		OffchainKeyring:              kb,
-		OnchainKeyring:               kb,
-		ContractTransmitter:          provider.ContractTransmitter(),
-		ContractConfigTracker:        provider.ContractConfigTracker(),
-		OffchainConfigDigester:       provider.OffchainConfigDigester(),
-	}
-
 	envVars, err := plugins.ParseEnvFile(plugEnv.Env.Get())
 	if err != nil {
 		return nil, fmt.Errorf("failed to parse median env file: %w", err)
 	}
-	if len(p.EnvVars) > 0 {
-		for k, v := range p.EnvVars {
+	if len(pCfg.EnvVars) > 0 {
+		for k, v := range pCfg.EnvVars {
 			envVars = append(envVars, k+"="+v)
 		}
 	}
 
-	pluginLggr := lggr.Named(p.PluginName).Named(spec.ContractID).Named(spec.GetID())
+	pluginLggr := lggr.Named(pCfg.PluginName).Named(spec.ContractID).Named(spec.GetID())
 	cmdFn, grpcOpts, err := d.cfg.RegisterLOOP(plugins.CmdConfig{
-		ID:  fmt.Sprintf("%s-%s-%s", p.PluginName, spec.ContractID, spec.GetID()),
+		ID:  fmt.Sprintf("%s-%s-%s", pCfg.PluginName, spec.ContractID, spec.GetID()),
 		Cmd: command,
 		Env: envVars,
 	})
@@ -616,9 +611,9 @@ func (d *Delegate) newServicesGenericPlugin(
 		//TODO: remove this workaround when the EVM relayer is running inside of an LOOPP
 		d.lggr.Info("provider is not a LOOPP provider, switching to provider server")
 
-		ps, err2 := relay.NewProviderServer(provider, types.OCR2PluginType(p.ProviderType), d.lggr)
+		ps, err2 := relay.NewProviderServer(provider, types.OCR2PluginType(pCfg.ProviderType), d.lggr)
 		if err2 != nil {
-			return nil, fmt.Errorf("cannot start EVM provider server: %s", err)
+			return nil, fmt.Errorf("cannot start EVM provider server: %s", err2)
 		}
 		providerClientConn, err2 = ps.GetConn()
 		if err2 != nil {
@@ -627,32 +622,82 @@ func (d *Delegate) newServicesGenericPlugin(
 		srvs = append(srvs, ps)
 	}
 
-	pc, err := json.Marshal(p.Config)
+	pc, err := json.Marshal(pCfg.Config)
 	if err != nil {
 		return nil, fmt.Errorf("cannot dump plugin config to string before sending to plugin: %s", err)
 	}
 
 	pluginConfig := types.ReportingPluginServiceConfig{
-		PluginName:    p.PluginName,
+		PluginName:    pCfg.PluginName,
 		Command:       command,
-		ProviderType:  p.ProviderType,
-		TelemetryType: p.TelemetryType,
+		ProviderType:  pCfg.ProviderType,
+		TelemetryType: pCfg.TelemetryType,
 		PluginConfig:  string(pc),
 	}
 
 	pr := generic.NewPipelineRunnerAdapter(pluginLggr, jb, d.pipelineRunner)
 	ta := generic.NewTelemetryAdapter(d.monitoringEndpointGen)
 
-	plugin := reportingplugins.NewLOOPPService(pluginLggr, grpcOpts, cmdFn, pluginConfig, providerClientConn, pr, ta, errorLog)
-	oracleArgs.ReportingPluginFactory = plugin
-	srvs = append(srvs, plugin)
+	oracleEndpoint := d.monitoringEndpointGen.GenMonitoringEndpoint(
+		rid.Network,
+		rid.ChainID,
+		spec.ContractID,
+		synchronization.TelemetryType(pCfg.TelemetryType),
+	)
 
-	oracle, err := libocr2.NewOracle(oracleArgs)
-	if err != nil {
-		return nil, err
+	switch pCfg.OCRVersion {
+	case 2:
+		plugin := reportingplugins.NewLOOPPService(pluginLggr, grpcOpts, cmdFn, pluginConfig, providerClientConn, pr, ta, errorLog)
+		oracleArgs := libocr2.OCR2OracleArgs{
+			BinaryNetworkEndpointFactory: d.peerWrapper.Peer2,
+			V2Bootstrappers:              bootstrapPeers,
+			Database:                     ocrDB,
+			LocalConfig:                  lc,
+			Logger:                       ocrLogger,
+			MonitoringEndpoint:           oracleEndpoint,
+			OffchainKeyring:              kb,
+			OnchainKeyring:               kb,
+			ContractTransmitter:          provider.ContractTransmitter(),
+			ContractConfigTracker:        provider.ContractConfigTracker(),
+			OffchainConfigDigester:       provider.OffchainConfigDigester(),
+		}
+		oracleArgs.ReportingPluginFactory = plugin
+		srvs = append(srvs, plugin)
+		oracle, err := libocr2.NewOracle(oracleArgs)
+		if err != nil {
+			return nil, err
+		}
+		srvs = append(srvs, job.NewServiceAdapter(oracle))
+
+	case 3:
+		//OCR3 with OCR2 OnchainKeyring and ContractTransmitter
+		plugin := ocr3.NewLOOPPService(pluginLggr, grpcOpts, cmdFn, pluginConfig, providerClientConn, pr, ta, errorLog, capabilitiesRegistry)
+		contractTransmitter := ocrcommon.NewOCR3ContractTransmitterAdapter(provider.ContractTransmitter())
+		oracleArgs := libocr2.OCR3OracleArgs[[]byte]{
+			BinaryNetworkEndpointFactory: d.peerWrapper.Peer2,
+			V2Bootstrappers:              bootstrapPeers,
+			ContractConfigTracker:        provider.ContractConfigTracker(),
+			ContractTransmitter:          contractTransmitter,
+			Database:                     ocrDB,
+			LocalConfig:                  lc,
+			Logger:                       ocrLogger,
+			MonitoringEndpoint:           oracleEndpoint,
+			OffchainConfigDigester:       provider.OffchainConfigDigester(),
+			OffchainKeyring:              kb,
+			OnchainKeyring:               ocrcommon.NewOCR3OnchainKeyringAdapter(kb),
+		}
+		oracleArgs.ReportingPluginFactory = plugin
+		srvs = append(srvs, plugin)
+		oracle, err := libocr2.NewOracle(oracleArgs)
+		if err != nil {
+			return nil, err
+		}
+		srvs = append(srvs, job.NewServiceAdapter(oracle))
+
+	default:
+		return nil, fmt.Errorf("unknown OCR version: %d", pCfg.OCRVersion)
 	}
 
-	srvs = append(srvs, job.NewServiceAdapter(oracle))
 	return srvs, nil
 }
 
@@ -739,6 +784,135 @@ func (d *Delegate) newServicesMercury(
 	}
 
 	return mercuryServices, err2
+}
+
+func (d *Delegate) newServicesLLO(
+	ctx context.Context,
+	lggr logger.SugaredLogger,
+	jb job.Job,
+	bootstrapPeers []commontypes.BootstrapperLocator,
+	kb ocr2key.KeyBundle,
+	ocrDB *db,
+	lc ocrtypes.LocalConfig,
+	ocrLogger commontypes.Logger,
+) ([]job.ServiceCtx, error) {
+	lggr = logger.Sugared(lggr.Named("LLO"))
+	spec := jb.OCR2OracleSpec
+	transmitterID := spec.TransmitterID.String
+	if len(transmitterID) != 64 {
+		return nil, errors.Errorf("ServicesForSpec: streams job type requires transmitter ID to be a 32-byte hex string, got: %q", transmitterID)
+	}
+	if _, err := hex.DecodeString(transmitterID); err != nil {
+		return nil, errors.Wrapf(err, "ServicesForSpec: streams job type requires transmitter ID to be a 32-byte hex string, got: %q", transmitterID)
+	}
+
+	rid, err := spec.RelayID()
+	if err != nil {
+		return nil, ErrJobSpecNoRelayer{Err: err, PluginName: "streams"}
+	}
+	if rid.Network != relay.EVM {
+		return nil, fmt.Errorf("streams services: expected EVM relayer got %s", rid.Network)
+	}
+	relayer, err := d.RelayGetter.Get(rid)
+	if err != nil {
+		return nil, ErrRelayNotEnabled{Err: err, Relay: spec.Relay, PluginName: "streams"}
+	}
+
+	provider, err2 := relayer.NewLLOProvider(ctx,
+		types.RelayArgs{
+			ExternalJobID: jb.ExternalJobID,
+			JobID:         jb.ID,
+			ContractID:    spec.ContractID,
+			New:           d.isNewlyCreatedJob,
+			RelayConfig:   spec.RelayConfig.Bytes(),
+			ProviderType:  string(spec.PluginType),
+		}, types.PluginArgs{
+			TransmitterID: transmitterID,
+			PluginConfig:  spec.PluginConfig.Bytes(),
+		})
+	if err2 != nil {
+		return nil, err2
+	}
+
+	var pluginCfg lloconfig.PluginConfig
+	if err = json.Unmarshal(spec.PluginConfig.Bytes(), &pluginCfg); err != nil {
+		return nil, err
+	}
+
+	kbm := make(map[llotypes.ReportFormat]llo.Key)
+	for rfStr, kbid := range pluginCfg.KeyBundleIDs {
+		k, err3 := d.ks.Get(kbid)
+		if err3 != nil {
+			return nil, fmt.Errorf("job %d (%s) specified key bundle ID %q for report format %s, but got error trying to load it: %w", jb.ID, jb.Name.ValueOrZero(), kbid, rfStr, err3)
+		}
+		rf, err4 := llotypes.ReportFormatFromString(rfStr)
+		if err4 != nil {
+			return nil, fmt.Errorf("job %d (%s) specified key bundle ID %q for report format %s, but it is not a recognized report format: %w", jb.ID, jb.Name.ValueOrZero(), kbid, rfStr, err4)
+		}
+		kbm[rf] = k
+	}
+	// NOTE: This is a bit messy because we assume chain type matches report
+	// format, and it may not in all cases. We don't yet know what report
+	// formats we need or how they correspond to chain types, so assume it's
+	// 1:1 for now but will change in future
+	//
+	// https://smartcontract-it.atlassian.net/browse/MERC-3722
+	for _, s := range chaintype.SupportedChainTypes {
+		rf, err3 := llotypes.ReportFormatFromString(string(s))
+		if err3 != nil {
+			return nil, fmt.Errorf("job %d (%s) has a chain type with no matching report format %s: %w", jb.ID, jb.Name.ValueOrZero(), s, err3)
+		}
+		if _, exists := kbm[rf]; !exists {
+			// Use the first if unspecified
+			kbs, err4 := d.ks.GetAllOfType(s)
+			if err4 != nil {
+				return nil, err4
+			}
+			if len(kbs) == 0 {
+				// unsupported key type
+				continue
+			} else if len(kbs) > 1 {
+				lggr.Debugf("Multiple on-chain signing keys found for report format %s, using the first", rf.String())
+			}
+			kbm[rf] = kbs[0]
+		}
+	}
+
+	// FIXME: This is a bit confusing because the OCR2 key bundle actually
+	// includes an EVM on-chain key... but LLO only uses the key bundle for the
+	// offchain keys and the suppoprted onchain keys are defined in the plugin
+	// config on the job spec instead.
+	// https://smartcontract-it.atlassian.net/browse/MERC-3594
+	lggr.Infof("Using on-chain signing keys for LLO job %d (%s): %v", jb.ID, jb.Name.ValueOrZero(), kbm)
+	kr := llo.NewOnchainKeyring(lggr, kbm)
+
+	cfg := llo.DelegateConfig{
+		Logger:   lggr,
+		Queryer:  pg.NewQ(d.db, lggr, d.cfg.Database()),
+		Runner:   d.pipelineRunner,
+		Registry: d.streamRegistry,
+
+		ChannelDefinitionCache: provider.ChannelDefinitionCache(),
+
+		BinaryNetworkEndpointFactory: d.peerWrapper.Peer2,
+		V2Bootstrappers:              bootstrapPeers,
+		ContractTransmitter:          provider.ContractTransmitter(),
+		ContractConfigTracker:        provider.ContractConfigTracker(),
+		Database:                     ocrDB,
+		LocalConfig:                  lc,
+		// TODO: Telemetry for llo
+		// https://smartcontract-it.atlassian.net/browse/MERC-3603
+		MonitoringEndpoint:     nil,
+		OffchainConfigDigester: provider.OffchainConfigDigester(),
+		OffchainKeyring:        kb,
+		OnchainKeyring:         kr,
+		OCRLogger:              ocrLogger,
+	}
+	oracle, err := llo.NewDelegate(cfg)
+	if err != nil {
+		return nil, err
+	}
+	return []job.ServiceCtx{provider, oracle}, nil
 }
 
 func (d *Delegate) newServicesMedian(